Simplify the structure of the CLI modules
