+++ src/cli.rs
@@ -5,17 +5,16 @@
 // you may not use this file except in compliance with the License.
 // A copy of the License has been included in the root of the repository.
 
-//! Functions to parse the command line.
+//! Types that represent a parsed command line, and functions to parse it.
 
 use std::str::FromStr;
 
-use crate::cli::arg_iter::{Arg, ArgIter};
-use crate::cli::command::{
-    Cmd, FormatOptions, FormatTarget, GlobalOptions, OutputFormat, OutputOptions, Target,
-};
+use crate::cli_utils::{match_option, parse_option, Arg, ArgIter};
 use crate::error::{Error, Result};
+use crate::loader::Loader;
 use crate::markup::{Markup, MarkupMode};
-use crate::pprint::{concat, Doc};
+use crate::pprint::{self, concat, Doc};
+use crate::source::DocId;
 
 const USAGE_MAIN: &str = r#"
 RCL -- A sane configuration language.
@@ -95,67 +94,102 @@ Options:
 See also --help for global options.
 "#;
 
-/// Helper macro to match option values and assign them to a field on match.
-///
-/// This also takes care of generating the errors for non-match cases.
-macro_rules! match_option {
-    {
-        $args_iter:ident: $option:expr,
-        $( $pat:literal => $val:expr , )+
-    } => {{
-        let mut err = vec![
-            "Expected ".into(),
-            Doc::from($option.to_string()).with_markup(Markup::Highlight),
-            " to be followed by one of ".into(),
-        ];
-        $(
-            err.push(Doc::from($pat).with_markup(Markup::Highlight));
-            err.push(", ".into());
-        )+
-        err.pop();
-        err.push(". See --help for usage.".into());
-        let err = Doc::Concat(err);
-
-        match $args_iter.next() {
-            Some(arg) => match arg.as_ref() {
-                $( Arg::Plain($pat) => $val, )+
-                _ => return Error::new(err).err(),
-            }
-            None => return Error::new(err).err(),
-        }
-    }};
+/// Options that apply to all subcommands.
+#[derive(Debug, Default, Eq, PartialEq)]
+pub struct GlobalOptions {
+    /// Whether and how to output color and other markup.
+    ///
+    /// We call it “markup” internally because it's more than just color, but
+    /// we call it `--color` on the command line because that is what most tools
+    /// call it.
+    pub markup: Option<MarkupMode>,
 }
 
-macro_rules! parse_option {
-    {
-        $args_iter:ident: $option:expr, $parser:expr
-    } => {
-        match $args_iter.next() {
-            Some(Arg::Plain(value)) => match $parser(&value[..]) {
-                Ok(r) => r,
-                Err(..) => {
-                    let err = concat! {
-                        "'"
-                        Doc::lines(&value).into_owned().with_markup(Markup::Highlight)
-                        "' is not valid for "
-                        Doc::from($option.to_string()).with_markup(Markup::Highlight)
-                        ". See --help for usage."
-                    };
-                    return Error::new(err).err();
-                }
-            }
-            _ => {
-                let err = concat! {
-                    "Expected a value after "
-                    Doc::from($option.to_string()).with_markup(Markup::Highlight)
-                    ". See --help for usage."
-                };
-                return Error::new(err).err();
-            }
-        }
-    };
+impl GlobalOptions {
+    /// Overwrite the relevant settings on the pprint config.
+    pub fn apply(&self, cfg: &mut pprint::Config) {
+        cfg.markup = self.markup.unwrap_or(cfg.markup);
+    }
 }
 
+/// The available output formats (JSON, RCL).
+#[derive(Debug, Default, Eq, PartialEq)]
+pub enum OutputFormat {
+    Json,
+    #[default]
+    Rcl,
+}
+
+/// Options for commands that output values.
+#[derive(Debug, Default, Eq, PartialEq)]
+pub struct OutputOptions {
+    /// The format to output in.
+    pub format: OutputFormat,
+}
+
+/// Options for commands that pretty-print their output.
+#[derive(Debug, Default, Eq, PartialEq)]
+pub struct FormatOptions {
+    /// Target width (number of columns) to try to not exceed.
+    pub width: Option<u32>,
+}
+
+impl FormatOptions {
+    /// Overwrite the relevant settings on the pprint config.
+    pub fn apply(&self, cfg: &mut pprint::Config) {
+        cfg.width = self.width.unwrap_or(cfg.width);
+    }
+}
+
+/// Input to act on.
+pub enum Target {
+    File(String),
+    Stdin,
+}
+
+impl Target {
+    /// Load the input into the loader.
+    pub fn load(&self, loader: &mut Loader) -> Result<DocId> {
+        match self {
+            Target::File(path) => loader.load_file(path),
+            Target::Stdin => loader.load_stdin(),
+        }
+    }
+}
+
+/// For the `fmt` command, which documents to format, and in what mode.
+pub enum FormatTarget {
+    Stdout { fname: Target },
+    InPlace { fnames: Vec<Target> },
+}
+
+/// The different subcommands supported by the main program.
+pub enum Cmd {
+    Evaluate {
+        output_opts: OutputOptions,
+        format_opts: FormatOptions,
+        fname: Target,
+    },
+    Query {
+        output_opts: OutputOptions,
+        format_opts: FormatOptions,
+        fname: Target,
+        query: String,
+    },
+    Format {
+        format_opts: FormatOptions,
+        target: FormatTarget,
+    },
+    Highlight {
+        fname: Target,
+    },
+    Help {
+        usage: &'static str,
+    },
+    Version,
+}
+
+/// Parse the command line.
 pub fn parse(args: Vec<String>) -> Result<(GlobalOptions, Cmd)> {
     let mut args = ArgIter::new(args);
 
@@ -320,7 +354,7 @@ pub fn parse(args: Vec<String>) -> Result<(GlobalOptions, Cmd)> {
     Ok((global_opts, result))
 }
 
-pub fn get_unique_target(mut targets: Vec<Target>) -> Result<Target> {
+fn get_unique_target(mut targets: Vec<Target>) -> Result<Target> {
     match targets.pop() {
         None => Error::new("Expected an input file. See --help for usage.").err(),
         Some(_) if !targets.is_empty() => {
