+++ src/lib.rs
@@ -10,13 +10,12 @@
 #![allow(clippy::len_without_is_empty)]
 #![allow(clippy::manual_range_contains)]
 
+mod cli_utils;
+mod platform_utils;
+
 pub mod abstraction;
 pub mod ast;
-pub mod cli {
-    pub mod arg_iter;
-    pub mod cli_parser;
-    pub mod command;
-}
+pub mod cli;
 pub mod cst;
 pub mod error;
 pub mod eval;
@@ -28,7 +27,6 @@ pub mod lexer;
 pub mod loader;
 pub mod markup;
 pub mod parser;
-pub mod platform_utils;
 pub mod pprint;
 pub mod runtime;
 pub mod source;
